Update JSON menu representation in mutations
